NDRS-591: update JSON helper types
