+++ node/src/components/event_stream_server/sse_server.rs
@@ -17,7 +17,8 @@ use crate::{
     components::CLIENT_API_VERSION,
     crypto::asymmetric_key::PublicKey,
     types::{
-        BlockHash, BlockHeader, DeployHash, ExecutionResult, FinalizedBlock, TimeDiff, Timestamp,
+        json_compatibility::ExecutionResult, BlockHash, BlockHeader, DeployHash, FinalizedBlock,
+        TimeDiff, Timestamp,
     },
 };
