+++ node/src/components/rpc_server/rpcs/chain.rs
@@ -2,7 +2,6 @@
 
 use std::str;
 
-use datasize::DataSize;
 use futures::{future::BoxFuture, FutureExt};
 use http::Response;
 use hyper::Body;
@@ -11,22 +10,15 @@ use serde::{Deserialize, Serialize};
 use tracing::info;
 use warp_json_rpc::Builder;
 
-use casper_types::auction::EraId;
-
 use super::{
     Error, ErrorCode, ReactorEventT, RpcRequest, RpcWithOptionalParams, RpcWithOptionalParamsExt,
 };
 use crate::{
     components::CLIENT_API_VERSION,
-    crypto::{
-        asymmetric_key::{PublicKey, Signature},
-        hash::Digest,
-    },
+    crypto::hash::Digest,
     effect::EffectBuilder,
     reactor::QueueKind,
-    types::{
-        json_compatibility::KeyValuePair, Block, BlockHash, BlockHeader, DeployHash, Timestamp,
-    },
+    types::{Block, BlockHash, JsonBlock},
 };
 
 /// Identifier for possible ways to retrieve a block.
@@ -36,111 +28,6 @@ pub enum BlockIdentifier {
     Hash(BlockHash),
     /// Identify and retrieve the block with its height.
     Height(u64),
-}
-
-type EraEnd = crate::components::consensus::EraEnd<PublicKey>;
-
-/// Represent the BTreeMap<PublicKey,u64> in an OpenRPC compatible format.
-#[derive(Clone, DataSize, Eq, PartialEq, Serialize, Deserialize, Debug)]
-pub struct JsonEraEnd {
-    equivocators: Vec<PublicKey>,
-    rewards: Vec<KeyValuePair<PublicKey, u64>>,
-}
-
-impl JsonEraEnd {
-    /// Create a single representation of EraEnd that can be serialized to an OpenRPC compatible
-    /// format.
-    fn new(equivocators: Vec<PublicKey>, rewards: Vec<KeyValuePair<PublicKey, u64>>) -> Self {
-        JsonEraEnd {
-            equivocators,
-            rewards,
-        }
-    }
-}
-
-impl From<&EraEnd> for JsonEraEnd {
-    fn from(era_end: &EraEnd) -> Self {
-        let rewards = era_end
-            .rewards
-            .iter()
-            .map(|(public_key, rewards)| KeyValuePair::new(*public_key, *rewards))
-            .collect();
-        JsonEraEnd::new(era_end.equivocators.clone(), rewards)
-    }
-}
-
-/// The header portion of a [`Block`](struct.Block.html).
-#[derive(Clone, DataSize, Eq, PartialEq, Serialize, Deserialize, Debug)]
-pub struct JsonBlockHeader {
-    parent_hash: BlockHash,
-    state_root_hash: Digest,
-    body_hash: Digest,
-    deploy_hashes: Vec<DeployHash>,
-    random_bit: bool,
-    accumulated_seed: Digest,
-    era_end: Option<JsonEraEnd>,
-    timestamp: Timestamp,
-    era_id: EraId,
-    height: u64,
-    proposer: PublicKey,
-}
-
-impl JsonBlockHeader {
-    /// The list of deploy hashes included in the block.
-    pub fn deploy_hashes(&self) -> &Vec<DeployHash> {
-        &self.deploy_hashes
-    }
-}
-
-impl From<BlockHeader> for JsonBlockHeader {
-    fn from(block_header: BlockHeader) -> Self {
-        JsonBlockHeader {
-            parent_hash: *block_header.parent_hash(),
-            state_root_hash: *block_header.state_root_hash(),
-            body_hash: *block_header.body_hash(),
-            deploy_hashes: block_header.deploy_hashes().clone(),
-            random_bit: block_header.random_bit(),
-            accumulated_seed: block_header.accumulated_seed(),
-            era_end: Some(block_header.era_end().unwrap().into()),
-            timestamp: block_header.timestamp(),
-            era_id: block_header.era_id().0,
-            height: block_header.height(),
-            proposer: *block_header.proposer(),
-        }
-    }
-}
-
-/// A proto-block after execution, with the resulting post-state-hash.  This is the core component
-/// of the Casper linear blockchain represented in an OpenRPC compatible manner.
-#[derive(DataSize, Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct JsonBlock {
-    hash: BlockHash,
-    header: JsonBlockHeader,
-    body: (), // TODO: implement body of block
-    proofs: Vec<Signature>,
-}
-
-impl JsonBlock {
-    fn new(hash: BlockHash, header: JsonBlockHeader, proofs: Vec<Signature>) -> Self {
-        JsonBlock {
-            hash,
-            header,
-            body: (),
-            proofs,
-        }
-    }
-
-    /// The deploy hashes included in this block.
-    pub fn deploy_hashes(&self) -> &Vec<DeployHash> {
-        self.header.deploy_hashes()
-    }
-}
-
-impl From<Block> for JsonBlock {
-    fn from(block: Block) -> Self {
-        let header: JsonBlockHeader = block.header().clone().into();
-        JsonBlock::new(*block.hash(), header, block.proofs().clone())
-    }
 }
 
 /// Params for "chain_get_block" RPC request.
