+++ node/src/types/deploy.rs
@@ -21,7 +21,7 @@ use casper_execution_engine::core::engine_state::{
 };
 use casper_types::bytesrepr::{self, FromBytes, ToBytes};
 
-use super::{BlockHash, ExecutionResult, Item, Tag, TimeDiff, Timestamp};
+use super::{json_compatibility::ExecutionResult, BlockHash, Item, Tag, TimeDiff, Timestamp};
 #[cfg(test)]
 use crate::testing::TestRng;
 use crate::{
